reworked variable types

+++ ROS Nodes/right_wheels.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 import rospy
-from std_msgs.msg import Int16
+from std_msgs.msg import Int32
 
 def callback(data):
     rospy.loginfo(data.data)
@@ -8,7 +8,7 @@ def callback(data):
 def right_wheels():
 
     rospy.init_node('right_wheels')
-    rospy.Subscriber('right_wheels', Int16, callback)
+    rospy.Subscriber('right_wheels', Int32, callback)
     rospy.spin()
 
 if __name__ == '__main__':
